#14580: Remove usage of noc_parameters.h from cq_commands.hpp
