+++ tt_metal/impl/dispatch/kernels/cq_prefetch.cpp
@@ -13,6 +13,17 @@
 #include "tt_metal/impl/dispatch/kernels/cq_common.hpp"
 #include "debug/dprint.h"
 
+#include "noc/noc_parameters.h" // PCIE_ALIGNMENT
+constexpr uint32_t CQ_PREFETCH_CMD_BARE_MIN_SIZE = PCIE_ALIGNMENT; // for NOC PCIe alignemnt
+struct CQPrefetchHToPrefetchDHeader_s {
+    uint32_t length;
+};
+typedef union {
+    struct CQPrefetchHToPrefetchDHeader_s header;
+    unsigned char padding[CQ_PREFETCH_CMD_BARE_MIN_SIZE];
+} CQPrefetchHToPrefetchDHeader;
+static_assert((sizeof(CQPrefetchHToPrefetchDHeader) & (CQ_PREFETCH_CMD_BARE_MIN_SIZE - 1)) == 0);
+
 typedef uint16_t prefetch_q_entry_type;
 
 constexpr uint32_t downstream_cb_base = get_compile_time_arg_val(0);
